+++ src/gods/poseidon_reload.lua
@@ -1,38 +1,50 @@
-function mod.PoseidonWrath (unit, functionArgs, triggerArgs)
-	-- If Hero doesnt have Ocean Swell, dont crash, just dont do it
-	if not HeroHasTrait("OmegaPoseidonProjectileBoon") then
+modutil.mod.Path.Override("CheckPoseidonFont", function (victim, triggerArgs)
+	local effectData = EffectData.AmplifyKnockbackEffect
+	if not triggerArgs.SourceProjectile or effectData.ProjectileNameBlacklistLookup[triggerArgs.SourceProjectile] then
 		return
 	end
+	if HeroHasTrait(gods.GetInternalBoonName("PoseidonWrathBoon")) then
+		mod.PoseidonWrath(victim, triggerArgs)
+	else
+		if RandomChance( effectData.Chance * GetTotalHeroTraitValue("LuckMultiplier", {IsMultiplier = true})) and CheckCooldown("PoseidonStatusFont", effectData.Cooldown) then
+			CreateProjectileFromUnit({ Name = effectData.ProjectileName, DestinationId = victim.ObjectId, Id = CurrentRun.Hero.ObjectId, FireFromTarget = true, DamageMultiplier = GetTotalHeroTraitValue("PoseidonFontMultiplier", { IsMultiplier = true }) })
+		end
+	end
+end)
 
+function mod.PoseidonWrath (victim, triggerArgs)
+	local effectData = EffectData.AmplifyKnockbackEffect
+	local omegaWaveChance = GetTotalHeroTraitValue("WaveChance")
 	local traitData = GetHeroTrait("OmegaPoseidonProjectileBoon")
-
-	-- This does nothing currently, but we can for example extract if red color must be used if trigger projectile is Ares one 
 	local dataProperties = GetProjectileProperty({ ProjectileId = triggerArgs.ProjectileId, Property = "DataProperties" })
-
 	local omegaPoseidonProjectile = 
 	{
-		Name = functionArgs.ProjectileName,
+		Name = "PoseidonOmegaWave",
 		Id = CurrentRun.Hero.ObjectId,
 		Angle = triggerArgs.ImpactAngle,
-		DestinationId = unit.ObjectId,
+		DestinationId = victim.ObjectId,
 		FireFromTarget = true,
-		DamageMultiplier = (traitData.OnWeaponFiredFunctions.FunctionArgs.DamageMultiplier or functionArgs.FallbackWeaponDamageMultiplier) * functionArgs.DamageMultiplier,
+		DamageMultiplier = (traitData.OnWeaponFiredFunctions.FunctionArgs.DamageMultiplier or 1.0),
 		DataProperties =
 		{
 			StartFx = dataProperties.StartFx,
-			ImpactVelocity = triggerArgs.ImpactVelocity or functionArgs.ImpactVelocity,
+			ImpactVelocity = triggerArgs.ImpactVelocity or 600, --Check ImpactVelocity in PoseidonWrath functionArgs
 			StartDelay = 0
 		},
 		ProjectileCap = 2,
 	}
-
-	local count = 1
-	for i=1, count do
-		CreateProjectileFromUnit(omegaPoseidonProjectile)
-		local doubleChance = GetTotalHeroTraitValue("DoubleOlympianProjectileChance") * GetTotalHeroTraitValue( "LuckMultiplier", { IsMultiplier = true })
-		if RandomChance(doubleChance) then
-			wait( GetTotalHeroTraitValue("DoubleOlympianProjectileInterval" ))
+	if CheckCooldown("PoseidonStatusFont", effectData.Cooldown) then
+		CreateProjectileFromUnit({ Name = effectData.ProjectileName, DestinationId = victim.ObjectId, Id = CurrentRun.Hero.ObjectId, FireFromTarget = true, DamageMultiplier = GetTotalHeroTraitValue("PoseidonFontMultiplier", { IsMultiplier = true }) })
+		if not HeroHasTrait("OmegaPoseidonProjectileBoon") then
+			return
+		end
+		if RandomChance( omegaWaveChance * GetTotalHeroTraitValue( "LuckMultiplier", { IsMultiplier = true }) ) then
 			CreateProjectileFromUnit(omegaPoseidonProjectile)
+			local doubleChance = GetTotalHeroTraitValue("DoubleOlympianProjectileChance") * GetTotalHeroTraitValue( "LuckMultiplier", { IsMultiplier = true })
+			if RandomChance(doubleChance) then
+				wait( GetTotalHeroTraitValue("DoubleOlympianProjectileInterval" ))
+				CreateProjectileFromUnit(omegaPoseidonProjectile)
+			end
 		end
 	end
 end